refactor: move main.ts to src directory
- Moved main.ts to src/main.ts for better project organization
- Updated import paths in src/settings.ts
- Updated esbuild config entry point
